* arts-1-3-31
* src/getopt.h: Fix compilation on Mac OS X Tiger.

* src/getopt.cc: Fix 64-bit compilation on Mac OS X Leopard.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5672 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/getopt.c
@@ -666,7 +666,7 @@ _getopt_internal (argc, argv, optstring, longopts, longind, long_only)
       /* Test all long options for either exact match
          or abbreviated matches.  */
       for (p = longopts, option_index = 0; p->name; p++, option_index++)
-        if (!strncmp (p->name, nextchar, (unsigned int)(nameend - nextchar)))
+        if (!strncmp (p->name, nextchar, (size_t)(nameend - nextchar)))
           {
             if ((unsigned int) (nameend - nextchar)
                 == (unsigned int) strlen (p->name))
@@ -855,7 +855,7 @@ _getopt_internal (argc, argv, optstring, longopts, longind, long_only)
         /* Test all long options for either exact match
            or abbreviated matches.  */
         for (p = longopts, option_index = 0; p->name; p++, option_index++)
-          if (!strncmp (p->name, nextchar, (unsigned int)(nameend - nextchar)))
+          if (!strncmp (p->name, nextchar, (size_t)(nameend - nextchar)))
             {
               if ((unsigned int) (nameend - nextchar) == strlen (p->name))
                 {
